requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
brotlipy==0.7.0
cachetools==5.3.1
ccimport==0.4.2
cumm-cu118==0.4.11
cycler==0.12.1
descartes==1.1.0
dropblock==0.3.0
einops==0.7.0
fire==0.5.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
imageio==2.34.0
importlib-metadata==6.8.0
joblib==1.3.2
kiwisolver==1.4.5
lark==1.1.7
lightning==2.1.0
lightning-utilities==0.9.0
llvmlite==0.41.1
loguru==0.7.2
Markdown==3.5
matplotlib==3.5.3
MinkowskiEngine==0.5.4
mkl-fft==1.3.1
mkl-service==2.4.0
multidict==6.0.4
ninja==1.11.1.1
numba==0.58.1
nuscenes-devkit==1.1.11
nvidia-ml-py==12.535.133
nvitop==1.3.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pccm==0.4.9
portalocker==2.8.2
protobuf==4.23.4
psutil==5.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.7
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-lightning==2.1.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
Shapely==1.8.5.post1
sheen==0.1.2
spconv-cu118==2.3.6
tabulate==0.9.0
tensorboard==2.15.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
